Split bashrc_add in separat files
If configuration is updated, sometimes it needed two logins to get
last configs, aliases and so on. Because first login it syncs the
configuration and load-functions, on second login, the new configuration
was used (valid for all functions in bashrc_add itself)

Now the loading of defaults is in a own file.
First the file is synced with git, then the new file gets sourced, so
the script can use the new configuration on first login